ducktape: move NodeCrash and BadLogLines to utils
